+++ resources/release-notes.md
@@ -1,3 +1,4 @@
 # Software
 
+- Improve management of compression in `mag.process.SignedInteger`
 - Add `RangeVariable` as option in `mag.process.Calibration`